requirements.txt
google-genai
python-dotenv
pandas
numpy
requests
